separate out modules and themes in build_distro.sh, add checks to make sure paths exist when building

+++ build_distro.sh
@@ -5,52 +5,64 @@ pull_git() {
     cd $BUILD_PATH/commons_profile
     git pull origin 7.x-3.x
 
-    cd $BUILD_PATH/repos
+    cd $BUILD_PATH/repos/modules
     for i in "${repos[@]}"; do
       echo $i
       cd $i
       git pull origin 7.x-3.x
       cd ..
     done
+    cd $BUILD_PATH/repos/themes/commons_origins
+    git pull origin 7.x-3.x
 }
 
 release_notes() {
   OUTPUT="Release Notes for $RELEASE"
   cd $BUILD_PATH/commons_profile
-  OUTPUT="$OUTPUT `drush rn --date $FROM_DATE $TO_DATE`"
+  OUTPUT="$OUTPUT \n `drush rn --date $FROM_DATE $TO_DATE`"
 
-  cd $BUILD_PATH/repos
+  cd $BUILD_PATH/repos/modules
   for i in "${repos[@]}"; do
     echo $i
     cd $i
-    OUTPUT="$OUTPUT `drush rn --date $FROM_DATE $TO_DATE`"
+    OUTPUT="$OUTPUT \n `drush rn --date $FROM_DATE $TO_DATE`"
     cd ..
   done
   echo $OUTPUT >> $BUILD_PATH/rn.txt
 }
 
 build_distro() {
-    cd $BUILD_PATH
-    rm -rf ./publish
-    # do we have the profile?
-    if [[ -d $BUILD_PATH/commons_profile ]]; then
-      if [[ -d $BUILD_PATH/repos ]]; then
-        drush make commons_profile/build-commons.make --no-cache --working-copy --prepare-install ./publish
-        rm -rf publish/profiles/commons/modules/contrib/commons*
-      else
-        drush make commons_profile/build-commons-dev.make --no-cache --working-copy --prepare-install ./publish
-        mkdir $BUILD_PATH/repos
-        mv publish/profiles/commons/modules/contrib/commons* repos/
-      fi
-      ln -sf $BUILD_PATH/repos/commons* publish/profiles/commons/modules/contrib/
-      chmod -R 777 publish/sites/default
-      # symlink the profile to our dev copy
-      rm -f publish/profiles/commons/*.*
-      rm -rf publish/profiles/commons/images
-      ln -s $BUILD_PATH/commons_profile/* publish/profiles/commons/
-      pull_git
+    if [[ -d $BUILD_PATH ]]; then
+        cd $BUILD_PATH
+        rm -rf ./publish
+        # do we have the profile?
+        if [[ -d $BUILD_PATH/commons_profile ]]; then
+          if [[ -d $BUILD_PATH/repos ]]; then
+            drush make commons_profile/build-commons.make --no-cache --working-copy --prepare-install ./publish
+            rm -rf publish/profiles/commons/modules/contrib/commons*
+            rm -rf publish/profiles/commons/themes/contrib/commons*
+          else
+            drush make commons_profile/build-commons-dev.make --no-cache --working-copy --prepare-install ./publish
+            mkdir $BUILD_PATH/repos
+            mkdir $BUILD_PATH/repos/modules
+            mkdir $BUILD_PATH/repos/themes
+            mv publish/profiles/commons/modules/contrib/commons* repos/modules/
+            mv publish/profiles/commons/themes/contrib/commons* repos/themes/
+          fi
+          ln -sf $BUILD_PATH/repos/modules/commons* publish/profiles/commons/modules/contrib/
+          ln -sf $BUILD_PATH/repos/themes/commons* publish/profiles/commons/themes/contrib/
+          chmod -R 777 publish/sites/default
+          # symlink the profile to our dev copy
+          rm -f publish/profiles/commons/*.*
+          rm -rf publish/profiles/commons/images
+          ln -s $BUILD_PATH/commons_profile/* publish/profiles/commons/
+          pull_git
+        else
+          git clone http://git.drupal.org/project/commons.git commons_profile
+          build_distro $BUILD_PATH
+        fi
     else
-      git clone http://git.drupal.org/project/commons.git commons_profile
+      mkdir $BUILD_PATH
       build_distro $BUILD_PATH
     fi
 }
